xarray: add xas_try_halve() to halve a multi-index entry.
It is a preparation patch for non-uniform folio split, which always split
a folio into half iteratively.

When a new xa_node is needed during the split, xas_try_halve() can try to
allocate one but no more. -ENOMEM will be return if a node cannot be
allocated. -EINVAL will be return if a sibling node is split or
cascade split happens, where two or more new nodes are needed, and these
are not supported by xas_try_halve().

Signed-off-by: Zi Yan <ziy@nvidia.com>